+++ packages/go_router_builder/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.41/Dart 3.11.
+
 ## 4.5.0
 
 - Detects routes that resolve to the same URL pattern. Routes are compared by